Added initial entity/entity collision
Now the player and enemies can collide and bounce off each other.
In addition, the player or enemies can die depending on who is
higher when they collide.  Unfortunately, STATUS_DEAD hasn't been
programmed yet, so the only result is that the player can't jump
when dead.  I need to add the sprites for being dead and for eggs,
and then program the player or enemy to fly off the screen and
despawn and the enemy to drop an egg upon losing a joust.