Improved collision
Now birds don't appear to get permanently stuck anymore and
collision feels overall more solid and consistent.  It's still
possible to get stuck in a platform, but I have a backup collision
test that forces the bird out of the platform if it does get stuck.
This was present before, but didn't always work because it actually
checked a different set of collision points.