Fixed screen wrap, updated stage to add platforms over lava
After I added platforms over the lava, it's possible to run
directly across the entire stage.  I was finding it irritating
to bounce off some invisible barrier on the sides of the stage.
I knew it was because I was attempting to read memory outside the
screenbuffer (or in other parts of the screenbuffer) when checking
for collision with getPixel, so I changed it so that we don't
check for collision when the destination is off the screen.  Now
the players wrap around the sides of the stage smoothly without
bouncing.