Added highscores and highscore entry
Now, if the player has a highscore, they are presented with a
screen to enter their initials, which are then inserted into the
highscore table.  Also, I have added the highscore table to the
title screen mode, so the highscores display when in the title.
I will soon add the actual title screen and alternate between the
title card itself and the highscores.

I can't seem to get the EEPROM data to save, but I'm not sure if
it's because the emulator doesn't support saving EEPROM for .hex
files dropped into it.  I may have to put off properly testing
actually saving the highscores between power cycles until I
hopefully get the actual hardware.  For now, though, the highscores
stay until the emulator is reset, so I know the highscore saving
code that doesn't deal with EEPROM works.