Added title screen with animation
Now when in MODE_TITLE, it alternates between the title card and
the highscores table.  Pressing A will start a new game.  The game
now starts in this mode, and it's nice and snazzy.

However, I'm not completely satisfied with the animated stripes.
I have changed the angles on each side from what they were in the
original game because I was unhappy with the vertically moving
stripes on the highscore screen.  The right side would align with
the top, but the left side wouldn't align with the top or bottom
due to the way the stripe angles work.  I changed them so that
none of them would line up, but I'm still not sure I'm satisfied
with the result.  It's a small detail, but I just want it to feel
right when starting the game.