Added spawn invincibility, shortened title sequence
Now a player is invincible after spawning until they hit a button.
This prevents the undesirable situation in which a player spawns
under an enemy and is instantly killed.  I made the player flash
rapidly while invincible.  I'm not sure I'm completely satisfied
with it, but it works well at indicating invincibility.

I also doubled up the rate at which the stripes on the title screen
are revealed and halved the time the title and highscores are each
shown, so the title and highscores cycle more frequently.