Added timeout on egg spawning in each wave
Now there is a time limit on each wave of one minute before eggs
stop spawning.  In the original game, a pterodactyl spawns to kill
the player if they dawdle too long, but it requires extreme
precision to defeat, and that would be hard on such a small screen.
Also, I don't want to punish players too much and would prefer if
they are able to make it to later waves.  However, to prevent
farming eggs (and points/lives), I added this limit.  I think a
minute per wave is pretty reasonable, but I can extend it if it
ever ends up being a problem.