Added enemy speed bonus after wave 40, fixed lava bug
Now enemies increase their speed every 10 waves after wave 40,
until they are travelling as fast as the player's max speed. Once
they get going that fast it's pretty chaotic, but that should be
very far into a game, and hopefully the player is skilled enough
by that point to handle it.

I also fixed a bug in which the player lost multiple lives after
touching lava, sometimes resulting in the value wrapping around
and giving the player over 250 lives.