Fixed highscore saving bug, enabled flapping with A or B
I had made a simple error with the way I was saving and loading
the highscores: I forgot to account for the offset of the magic
number I put at the start of the save data, so it overwrote it
and then refused to load the data since the magic number was
missing.

I also added the ability to flap with B as well as A.  Some players
might prefer to flap with one button or the other.